requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Authentication
PyJWT==2.8.0
python-multipart==0.0.6

# Database
supabase==2.3.0
postgrest==0.13.0

# Data processing
numpy==1.26.2
pandas==2.1.3

# HTTP requests
httpx==0.25.2
requests==2.31.0

# Utilities
python-dateutil==2.8.2
